Isolate the entire farm and statistics code
